Move file classes to language package
