+++ src/main/java/software/amazon/smithy/plugin/language/SmithyIcons.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.smithy.plugin;
+package software.amazon.smithy.plugin.language;
 
 import com.intellij.openapi.util.IconLoader;
 import javax.swing.Icon;
